requirements.txt
pandas>=1.5.0
openpyxl>=3.0.0
requests>=2.26.0
pyyaml>=6.0.0
nltk>=3.8.1
inflect>=7.0.0
tqdm>=4.66.1 